added another seed event and adjusted date types

+++ data/migrations/20200123231512_events.js
@@ -5,9 +5,9 @@ exports.up = function (knex) {
         table.timestamps(true, true);
 
         table.string('eventname').notNullable();
-        table.integer('date')
-        table.integer('starttime')
-        table.integer('endtime')
+        table.bigInteger('date')
+        table.bigInteger('starttime')
+        table.bigInteger('endtime')
         table.text('details')
         table.string('imageLink');
